requirements.txt
django
djangorestframework
apsceduler